fix: resolve all compilation errors and launch desktop app successfully
- Fixed final 2 Send trait errors by restructuring database transactions
- Updated Dioxus syntax for version 0.5 compatibility
- Moved CSS styles to const to avoid rsx! macro parsing issues
- Fixed all signal access patterns (read/write methods)
- Desktop app now compiles and runs successfully

✅ All compilation errors resolved (320 → 0)
✅ Desktop app launches with VS Code-style interface

+++ src/planning/risk_analyzer.rs
@@ -612,7 +612,7 @@ impl RiskAnalyzer {
 
     fn calculate_mitigation_cost(&self, risk: &Risk, template: &MitigationTemplate) -> f64 {
         let base_cost = 1000.0; // Base cost per mitigation
-        base_cost * template.cost_factor * (risk.probability as f64)
+        base_cost * (template.cost_factor as f64) * (risk.probability as f64)
     }
 
     fn calculate_mitigation_time(&self, risk: &Risk, template: &MitigationTemplate) -> Duration {
